CODAP-46 address PR comments
