+++ v3/src/components/app.tsx
@@ -1,11 +1,12 @@
 import { useDisclosure } from "@chakra-ui/react"
 import { observer } from "mobx-react-lite"
 import React, { useCallback, useEffect, useState } from "react"
-import { ErrorBoundary, FallbackProps } from "react-error-boundary"
+import { ErrorBoundary } from "react-error-boundary"
 import { CfmContext } from "../hooks/use-cfm-context"
 import { DocumentContentContext } from "../hooks/use-document-content"
 import {useDropHandler} from "../hooks/use-drop-handler"
 import { useKeyStates } from "../hooks/use-key-states"
+import { useUncaughtErrorHandler } from "../hooks/use-uncaught-error-handler"
 import { useCloudFileManager } from "../lib/cfm/use-cloud-file-manager"
 import { CodapDndContext } from "../lib/dnd-kit/codap-dnd-context"
 import { logStringifiedObjectMessage } from "../lib/log-message"
@@ -25,7 +26,6 @@ import { getSharedModelManager } from "../models/tiles/tile-environment"
 import { registerTileTypes } from "../register-tile-types"
 import { importSample, sampleData } from "../sample-data"
 import { urlParams } from "../utilities/url-params"
-import { t } from "../utilities/translation/translate"
 import { kCodapAppElementId, kUserEntryDropOverlay } from "./constants"
 import { Container } from "./container/container"
 import { MenuBar, kMenuBarElementId } from "./menu-bar/menu-bar"
@@ -173,56 +173,7 @@ export const App = observer(function App() {
     initialize()
   }, [cfmReadyPromise, onCloseUserEntry, onOpenUserEntry])
 
-  const userCanSaveCopy = useCallback(() => {
-    // TODO: There are other cases where the user can't save a copy
-    return !cfm.appOptions.hideMenuBar
-  }, [cfm.appOptions.hideMenuBar])
-
-  const fallbackRender = useCallback(({ error }: FallbackProps) => {
-    errorBoundaryErrors.push({ time: Date.now(), message: error.message })
-
-    // Notes:
-    // 1. If we are running embedded in the Activity Player or some other
-    // system that does autosaving, the file menu may be disabled so
-    // the user can't save a copy of the document.
-    // 2. We are not using the resetErrorBoundary function provided to the
-    // fallbackRender function because we are basically asking them to
-    // reload the page anyhow, so giving them another option to try
-    // to render the document again would just be confusing.
-    return (
-      <div className="document-render-error">
-        <h1>{t("DG.mainPage.documentDisplayError.title")}</h1>
-        <p>
-          {[
-            userCanSaveCopy()
-              ? t("DG.mainPage.documentDisplayError.suggestionWithSave")
-              : t("DG.mainPage.documentDisplayError.suggestion"),
-            " ",
-            t("DG.mainPage.documentDisplayError.reason")
-          ]}
-        </p>
-        <h2>{t("DG.mainPage.documentDisplayError.detailsTitle")}</h2>
-        <p>{error.message}</p>
-      </div>
-    )
-  }, [userCanSaveCopy])
-
-  useEffect(() => {
-    const showErrorDialog = (error: ErrorEvent) => {
-      const dialogMessage = [
-        t("DG.mainPage.errorDialog.description"),
-        userCanSaveCopy()
-          ? t("DG.mainPage.errorDialog.suggestionWithSave")
-          : t("DG.mainPage.errorDialog.suggestion"),
-        t("DG.mainPage.errorDialog.errorDetails", {vars: [error.message]})
-      ].join(" ")
-      appState.alert(dialogMessage, t("DG.mainPage.errorDialog.title"))
-    }
-    window.addEventListener("error", showErrorDialog)
-    return () => {
-      window.removeEventListener("error", showErrorDialog)
-    }
-  }, [userCanSaveCopy])
+  const { fallbackRender } = useUncaughtErrorHandler(cfm)
 
   return (
     <CodapDndContext>
